pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["sajad <fallahdoostsajad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
faker = "^25.8.0"
tqdm = "^4.66.4"
mimesis = "^17.0.0"
pytz = "^2024.1"
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
